consolidate locale scripts
